Added Attendence summary component
